+++ index.html
@@ -7,7 +7,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
   <title>Кекстаграм</title>
 </head>
 
@@ -228,6 +227,6 @@
       </div>
     </section>
   </template>
-
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
