+++ js/const.js
@@ -1,0 +1,33 @@
+const NAMES = [
+  "Анастасия",
+  "Илья",
+  "Ксения",
+  "Сергей",
+  "Анна",
+  "Илья",
+  "Никита",
+  "Алина",
+  "Саша",
+  "Мария",
+];
+
+const MESSAGE = [
+  "Всё отлично!",
+  "В целом всё неплохо. Но не всё.",
+  "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
+  "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
+  "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
+  "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!",
+];
+
+const DESCRIPTION = [
+  "Какое сегодня чудесное утро!",
+  "Отдыхаю с друзьями.",
+  "Всем доброе утро!",
+  "Оцените от 1 до 10",
+  "Сегодня на рабочем вайбе",
+  "Эстетика",
+  "На прогулке",
+];
+
+export {NAMES, MESSAGE, DESCRIPTION};
